backend/requirements.txt
fastapi
uvicorn
sqlalchemy
ortools
google-genai
pydantic
firebase-admin
python-dotenv
